p00040002/train_requirements.txt
scikit-learn==0.20.1
pandas==1.1.5
Flask==1.1.2
requests==2.25.0
h2o==3.32.1.3
argparse==1.4.0
h2o==3.32.1.3
psycopg2-binary
matplotlib
seaborn